+++ src/main/java/br/mds/inti/config/SecurityConfig.java
@@ -16,8 +16,7 @@ import org.springframework.security.crypto.scrypt.SCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
-import br.mds.inti.auth.JwtAuthenticationFilter;
-import br.mds.inti.auth.UserDetailsServiceImpl;
+import br.mds.inti.service.UserDetailsServiceImpl;
 
 @Configuration
 @EnableMethodSecurity
