+++ src/main/java/br/mds/inti/model/Follow.java
@@ -1,8 +1,8 @@
-package br.mds.inti.models;
+package br.mds.inti.model;
 
 import java.time.Instant;
 
-import br.mds.inti.models.pk.EventParticipantPK;
+import br.mds.inti.model.pk.FollowsPK;
 import jakarta.persistence.Column;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
@@ -20,22 +20,23 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @EqualsAndHashCode
 @Entity
-@Table(name = "event_participants")
-public class EventParticipant {
+@Table(name = "follows")
+public class Follow {
 
     @EmbeddedId
-    private EventParticipantPK id;
+    private FollowsPK id;
 
     @ManyToOne
-    @MapsId("profileId")
-    @JoinColumn(name = "profile_id")
-    private Profile profile;
+    @JoinColumn(name = "follower_profile_id", nullable = false)
+    @MapsId("followerProfileId")
+    private Profile followerProfile;
 
     @ManyToOne
-    @MapsId("eventId") // conex
-    @JoinColumn(name = "event_id")
-    private Event event;
+    @JoinColumn(name = "following_profile_id", nullable = false)
+    @MapsId("followingProfileId")
+    private Profile followingProfile;
 
     @Column(name = "created_at", nullable = false)
     private Instant createdAt;
+
 }
