+++ src/main/java/br/mds/inti/model/Like.java
@@ -1,15 +1,14 @@
-package br.mds.inti.models;
+package br.mds.inti.model;
 
 import java.time.Instant;
-import java.util.UUID;
 
+import br.mds.inti.model.pk.LikePk;
 import jakarta.persistence.Column;
+import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.MapsId;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -21,23 +20,20 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @EqualsAndHashCode
 @Entity
-@Table(name = "memberships")
-public class Membership {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
+@Table(name = "likes")
+public class Like {
+    @EmbeddedId
+    private LikePk id;
 
     @ManyToOne
-    @JoinColumn(name = "profile_id", nullable = false)
+    @JoinColumn(name = "user_id", nullable = true)
+    @MapsId("userId")
     private Profile profile;
 
     @ManyToOne
-    @JoinColumn(name = "organization_id", nullable = false)
-    private Profile organization;
-
-    @Column(name = "role", nullable = true)
-    private String role;
+    @JoinColumn(name = "post_id", nullable = true)
+    @MapsId("postId")
+    private Post post;
 
     @Column(name = "created_at", nullable = false)
     private Instant createdAt;
