+++ src/main/java/br/mds/inti/model/Post.java
@@ -1,7 +1,8 @@
-package br.mds.inti.models;
+package br.mds.inti.model;
 
-import java.math.BigDecimal;
 import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.UUID;
 
 import jakarta.persistence.Column;
@@ -11,6 +12,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -22,9 +24,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @EqualsAndHashCode
 @Entity
-@Table(name = "artist_products")
-public class ArtistProducts {
-
+@Table(name = "posts")
+public class Post {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
@@ -33,19 +34,21 @@ public class ArtistProducts {
     @JoinColumn(name = "profile_id", nullable = false)
     private Profile profile;
 
-    @Column(name = "title", nullable = false, length = 255)
-    private String title;
-
     @Column(name = "img_link", nullable = true, length = 255)
     private String imgLink;
 
-    @Column(name = "price", nullable = false, precision = 10, scale = 2)
-    private BigDecimal price;
+    @Column(name = "description", nullable = true)
+    private String description;
+
+    @Column(name = "likes_count", nullable = true)
+    private Integer likesCount;
 
     @Column(name = "created_at", nullable = false)
     private Instant createdAt;
 
-    @Column(name = "deleted_at", nullable = true)
-    private Instant deletedAt;
+    @OneToMany(mappedBy = "post")
+    private List<Like> likes = new ArrayList<>();
 
+    @OneToMany(mappedBy = "postId")
+    private List<Shared> shareds = new ArrayList<>();
 }
