+++ src/main/java/br/mds/inti/model/dto/auth/LoginRequest.java
@@ -1,4 +1,4 @@
-package br.mds.inti.auth.dto;
+package br.mds.inti.model.dto.auth;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
