+++ src/main/java/br/mds/inti/model/dto/auth/RegisterRequest.java
@@ -1,6 +1,6 @@
-package br.mds.inti.auth.dto;
+package br.mds.inti.model.dto.auth;
 
-import br.mds.inti.models.ENUM.ProfileType;
+import br.mds.inti.model.enums.ProfileType;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
