+++ src/main/java/br/mds/inti/repositories/ProfileRepository.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import br.mds.inti.models.Profile;
+import br.mds.inti.model.Profile;
 
 public interface ProfileRepository extends JpaRepository<Profile, UUID> {
     Optional<Profile> findByEmail(String email);
