+++ src/main/java/br/mds/inti/service/JwtService.java
@@ -1,4 +1,4 @@
-package br.mds.inti.auth;
+package br.mds.inti.service;
 
 import java.util.Date;
 
@@ -10,6 +10,7 @@ import com.auth0.jwt.algorithms.Algorithm;
 @Service
 public class JwtService {
 
+    // todo: secret deve ser salvo em um config map externo e puxado do application.properties
     private static final String secret = "arranha-ceu";
     private long expirationMillis = 1000L * 60 * 60 * 24 * 30;
     private Date expiresAt = new Date(System.currentTimeMillis() + expirationMillis);
